Implement working fix for repeated server selection
Server selection in ServerIndexItem was coupled with a fetch for server
show action and subsequent categories and channels info. However it
did not re-fetch the same server if the route pointed to the same server.
Tentative fix uses router history to go back to the previous path if the
same server was selected.