Setup frontend ActionCable subscriptions to channel
At the mounted ChannelDetail, subscriptions were made for the current
user and a message form component apart of the ChannelDetail. The
subscribed channel is then passed messages from the form to the
RoomChannel backend for model creation and broadcasting back to
any subscribed users of the channel. When ChannelDetail unmounts,
the subscription is cleaned up to avoid receiving messages from the
wrong chat channel. Extra seeds added for testing.